move defaultDerivePresenceState into own file
