+++ packages/editor/src/lib/config/TldrawEditorConfig.tsx
@@ -31,7 +31,7 @@ import {
 	defineMigrations,
 } from '@tldraw/tlstore'
 import { T } from '@tldraw/tlvalidate'
-import { Signal, computed } from 'signia'
+import { Signal } from 'signia'
 import { TLArrowShapeDef } from '../app/shapeutils/TLArrowUtil/TLArrowUtil'
 import { TLBookmarkShapeDef } from '../app/shapeutils/TLBookmarkUtil/TLBookmarkUtil'
 import { TLDrawShapeDef } from '../app/shapeutils/TLDrawUtil/TLDrawUtil'
@@ -46,6 +46,7 @@ import { TLTextShapeDef } from '../app/shapeutils/TLTextUtil/TLTextUtil'
 import { TLVideoShapeDef } from '../app/shapeutils/TLVideoUtil/TLVideoUtil'
 import { StateNodeConstructor } from '../app/statechart/StateNode'
 import { TLShapeDef, TLUnknownShapeDef } from './TLShapeDefinition'
+import { defaultDerivePresenceState } from './defaultDerivePresenceState'
 
 const CORE_SHAPE_DEFS = () =>
 	[
@@ -162,58 +163,4 @@ export class TldrawEditorConfig {
 			},
 		})
 	}
-}
-
-/** @internal */
-export const defaultDerivePresenceState = (store: TLStore): Signal<TLInstancePresence | null> => {
-	const $instance = store.query.record('instance', () => ({
-		id: { eq: store.props.instanceId },
-	}))
-	const $user = store.query.record('user', () => ({ id: { eq: store.props.userId } }))
-	const $userPresence = store.query.record('user_presence', () => ({
-		userId: { eq: store.props.userId },
-	}))
-	const $pageState = store.query.record('instance_page_state', () => ({
-		instanceId: { eq: store.props.instanceId },
-		pageId: { eq: $instance.value?.currentPageId ?? ('' as any) },
-	}))
-	const $camera = store.query.record('camera', () => ({
-		id: { eq: $pageState.value?.cameraId ?? ('' as any) },
-	}))
-	return computed('instancePresence', () => {
-		const pageState = $pageState.value
-		const instance = $instance.value
-		const user = $user.value
-		const userPresence = $userPresence.value
-		const camera = $camera.value
-		if (!pageState || !instance || !user || !userPresence || !camera) {
-			return null
-		}
-
-		return TLInstancePresence.create({
-			id: TLInstancePresence.createCustomId(store.props.instanceId),
-			instanceId: store.props.instanceId,
-			selectedIds: pageState.selectedIds,
-			brush: instance.brush,
-			scribble: instance.scribble,
-			userId: store.props.userId,
-			userName: user.name,
-			followingUserId: instance.followingUserId,
-			camera: {
-				x: camera.x,
-				y: camera.y,
-				z: camera.z,
-			},
-			color: userPresence.color,
-			currentPageId: instance.currentPageId,
-			cursor: {
-				x: userPresence.cursor.x,
-				y: userPresence.cursor.y,
-				rotation: instance.cursor.rotation,
-				type: instance.cursor.type,
-			},
-			lastActivityTimestamp: userPresence.lastActivityTimestamp,
-			screenBounds: instance.screenBounds,
-		})
-	})
 }
